requirements.txt
aiohttp==3.7.4.post0
altair==4.1.0
anyio==3.3.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.0
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.68.0
gitdb==4.0.7
GitPython==3.1.18
gunicorn==20.1.0
h11==0.12.0
idna==3.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.2
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkg_resources==0.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
requests-unixsocket==0.2.0
rich==10.7.0
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
starlette==0.14.2
streamlit==0.86.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
twitchAPI==2.3.2
twitchio==2.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
uvicorn==0.14.0
validators==0.18.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.1
websockets==9.1
widgetsnbextension==3.5.1
yarl==1.6.3
